Refactor: Move SplashScreenTestTags to designsystem module
Moves `SplashScreenTestTags` from `core/model` to `core/designsystem/component` to colocate it with the `SplashScreen` composable.
Updates imports and adds test tags to elements within `SplashScreen` for UI testing.
