+++ app/src/jvmTest/kotlin/com/mshdabiola/kmtemplate/KmtAppTest.kt
@@ -43,6 +43,7 @@ import com.mshdabiola.detail.navigation.navigateToDetail
 import com.mshdabiola.kmtemplate.ui.KmtAppState
 import com.mshdabiola.kmtemplate.ui.KmtAppTestTags
 import com.mshdabiola.designsystem.component.SplashScreen
+import com.mshdabiola.designsystem.component.SplashScreenTestTags
 import com.mshdabiola.kmtemplate.ui.rememberKmtAppState
 import com.mshdabiola.kmtemplate.util.KoinTestRule
 import com.mshdabiola.kmtemplate.util.TestLifecycleOwner
@@ -51,7 +52,6 @@ import com.mshdabiola.main.navigation.Main
 import com.mshdabiola.model.testtag.DetailScreenTestTags
 import com.mshdabiola.model.testtag.MainScreenTestTags
 import com.mshdabiola.model.testtag.SettingScreenTestTags
-import com.mshdabiola.model.testtag.SplashScreenTestTags
 import com.mshdabiola.setting.navigation.Setting
 import com.mshdabiola.setting.settingModule
 import com.mshdabiola.testing.fake.testDataModule
