+++ core/designsystem/src/commonMain/kotlin/com/mshdabiola/designsystem/component/SplashScreenTestTags.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mshdabiola.model.testtag
+package com.mshdabiola.designsystem.component
 
 // Test Tags Object for SplashScreen
 object SplashScreenTestTags {
